+++ R/cc_coun.R
@@ -81,9 +81,7 @@ cc_coun <- function(x,
       warning("no projection information for reference found, 
               assuming '+proj=longlat +datum=WGS84 +no_defs 
               +ellps=WGS84 +towgs84=0,0,0'")
-    }else if(sp::proj4string(ref) == wgs84){
-      sp::proj4string(ref) <- ""
-    }else{
+    }else if(sp::proj4string(ref) != wgs84){
       ref <- sp::spTransform(ref, sp::CRS(wgs84))
       warning("reprojecting reference to '+proj=longlat +datum=WGS84 
               +no_defs +ellps=WGS84 +towgs84=0,0,0'")