+++ R/cc_inst.R
@@ -17,7 +17,11 @@
 #' Default = \dQuote{decimallatitude}.
 #' @param buffer numerical. The buffer around each province or country
 #' centroid, where records should be flagged as problematic, in decimal
-#' degrees.  Default = 0.001 (= ~ 100m).
+#' degrees.  Default = 100m.
+#' @param geod logical. If TRUE the radius around each institution is calculated
+#' based on a sphere, buffer is in meters and independent of latitude. If FALSE
+#' the radius is calculated assuming planar coordinates and varies slightly with latitude,
+#' in this case buffer is in degrees. DEfault = T.
 #' @param ref a SpatialPointsDataframe. Providing the geographic gazetteer. Can
 #' be any SpatialPointsDataframe, but the structure must be identical to
 #' \code{\link{institutions}}.  Default = \code{\link{institutions}}
@@ -40,12 +44,14 @@
 #' cc_inst(x, value = "flagged", buffer = 5)#large buffer for demonstration
 #' 
 #' @export
+#' @importFrom geosphere destPoint
 #' @importFrom raster crop extent
-#' @importFrom sp SpatialPoints
+#' @importFrom sp coordinates CRS disaggregate over Polygon Polygons SpatialPolygons SpatialPoints
 cc_inst <- function(x, 
                     lon = "decimallongitude", 
                     lat = "decimallatitude", 
-                    buffer = 0.001,
+                    buffer = 100,
+                    geod = TRUE,
                     ref = NULL, 
                     value = "clean", 
                     verbose = TRUE) {
@@ -56,8 +62,17 @@ cc_inst <- function(x,
   if (verbose) {
     message("Testing biodiversity institutions")
   }
+  
+  if(buffer > 10 & !geod){
+    warnings("Using large buffer check 'geod'")
+  }
+  if(buffer < 100 & geod){
+    warnings("Using small buffer check 'geod'")
+  }
+  # set default projection
+  wgs84 <- "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"
 
-  dat <- sp::SpatialPoints(x[, c(lon, lat)])
+  dat <- sp::SpatialPoints(x[, c(lon, lat)], proj4string = sp::CRS(wgs84))
 
   # prepare reference dataset
   if (is.null(ref)) {
@@ -68,7 +83,8 @@ cc_inst <- function(x,
 
   # subset of testdatset according to limits
   ref <- raster::crop(
-    sp::SpatialPoints(ref[, c("decimallongitude", "decimallatitude")]),
+    sp::SpatialPoints(ref[, c("decimallongitude", "decimallatitude")], 
+                      proj4string = sp::CRS(wgs84)),
     limits
   )
 
@@ -76,8 +92,31 @@ cc_inst <- function(x,
   if (is.null(ref)) {
     out <- rep(TRUE, nrow(x))
   } else {
-    ref <- rgeos::gBuffer(ref, width = buffer, byid = TRUE)
-    out <- is.na(sp::over(x = dat, y = ref))
+    if(geod){
+      # credits to https://seethedatablog.wordpress.com/2017/08/03/euclidean-vs-geodesic-buffering-in-r/
+      dg <- seq(from = 0, to = 360, by = 5)
+      
+      buff_XY <- geosphere::destPoint(p = sp::coordinates(ref), 
+                                      b = rep(dg, each = length(ref)), 
+                                      d = buffer)
+      
+      id <- rep(1:length(ref), times = length(dg))
+      
+      
+      lst <- split(data.frame(buff_XY), f = id)
+      
+      # Make SpatialPolygons out of the list of coordinates
+      poly   <- lapply(lst, sp::Polygon, hole = FALSE)
+      polys  <- lapply(list(poly), sp::Polygons, ID = NA)
+      spolys <- sp::SpatialPolygons(Srl = polys, proj4string = CRS(wgs84))
+      ref <- sp::disaggregate(spolys)
+
+      #point in polygon test
+      out <- is.na(sp::over(x = dat, y = ref))
+    }else{
+      ref <- rgeos::gBuffer(ref, width = buffer, byid = TRUE)
+      out <- is.na(sp::over(x = dat, y = ref))
+    }
   }
 
   # create output based on value argument