+++ NEWS.md
@@ -1,3 +1,14 @@
+
+CoordinateCleaner 2.0-2 (2018-08-20)
+=========================
+
+### MAJOR IMPROVEMENTS
+
+  * sampling correction into cc_outl
+  * tutorial on outlier detection on the bookdown documentation
+  * subsampling heuristic in cc_outl
+  
+
 CoordinateCleaner 2.0-1 (2018-06-08)
 =========================